Break dependency cycles
