+++ pom.xml
@@ -577,4 +577,3 @@
   </dependencyManagement>
 
 </project>
-
