+++ shogun-boot/src/test/java/de/terrestris/shogun/boot/config/JacksonConfigTest.java
@@ -60,7 +60,7 @@ public class JacksonConfigTest {
                 hasJtsModuleRegistered = true;
             }
         }
-        Assertions.assertTrue(hasJtsModuleRegistered, "JTS Module is not registrered in ObjectMapper");
+        Assertions.assertTrue(hasJtsModuleRegistered, "JTS Module is not registered in ObjectMapper");
     }
 
     @Test
