+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/security/SecurityContextUtil.java
@@ -16,25 +16,17 @@
  */
 package de.terrestris.shogun.lib.security;
 
-import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.repository.UserRepository;
-import de.terrestris.shogun.lib.service.security.provider.UserProviderService;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
-import org.keycloak.KeycloakPrincipal;
-import org.keycloak.KeycloakSecurityContext;
-import org.keycloak.representations.AccessToken;
-import org.keycloak.representations.IDToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Component
 @Log4j2
@@ -43,25 +35,6 @@ public class SecurityContextUtil {
     @Autowired
     protected UserRepository userRepository;
 
-    @Autowired
-    private UserProviderService userProviderService;
-
-    @Transactional(readOnly = true)
-    public Optional<User> getUserBySession() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String keycloakUserId = SecurityContextUtil.getKeycloakUserIdFromAuthentication(authentication);
-
-        if (StringUtils.isEmpty(keycloakUserId)) {
-            return Optional.empty();
-        }
-
-        Optional<User> user = userRepository.findByKeycloakId(keycloakUserId);
-
-        user.ifPresent(value -> userProviderService.setTransientRepresentations(value));
-
-        return user;
-    }
-
     /**
      *
      * @return
@@ -69,49 +42,6 @@ public class SecurityContextUtil {
     public List<GrantedAuthority> getGrantedAuthorities() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         return new ArrayList<>(authentication.getAuthorities());
-    }
-
-    /**
-     * Returns the current user object from the database.
-     *
-     * @param authentication
-     * @return
-     */
-    public Optional<User> getUserFromAuthentication(Authentication authentication) {
-        final Object principal = authentication.getPrincipal();
-        if (!(principal instanceof KeycloakPrincipal)) {
-            return Optional.empty();
-        }
-        // get user info from authentication object
-        String keycloakUserId = getKeycloakUserIdFromAuthentication(authentication);
-        return userRepository.findByKeycloakId(keycloakUserId);
-    }
-
-    /**
-     * Return keycloak user id from {@link Authentication} object
-     *   - from {@link IDToken}
-     *   - from {@link org.keycloak.Token}
-     * @param authentication The Spring security authentication
-     * @return The keycloak user id token
-     */
-    public static String getKeycloakUserIdFromAuthentication(Authentication authentication) {
-        if (authentication.getPrincipal() instanceof KeycloakPrincipal) {
-            KeycloakPrincipal<KeycloakSecurityContext> keycloakPrincipal = (KeycloakPrincipal<KeycloakSecurityContext>) authentication.getPrincipal();
-            KeycloakSecurityContext keycloakSecurityContext = keycloakPrincipal.getKeycloakSecurityContext();
-            IDToken idToken = keycloakSecurityContext.getIdToken();
-            String keycloakUserId;
-
-            if (idToken != null) {
-                keycloakUserId = idToken.getSubject();
-            } else {
-                AccessToken accessToken = keycloakSecurityContext.getToken();
-                keycloakUserId = accessToken.getSubject();
-            }
-
-            return keycloakUserId;
-        } else {
-            return null;
-        }
     }
 
     /**
