+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/security/permission/UserClassPermissionService.java
@@ -23,24 +23,24 @@ import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
 import de.terrestris.shogun.lib.model.security.permission.UserClassPermission;
 import de.terrestris.shogun.lib.repository.security.permission.PermissionCollectionRepository;
 import de.terrestris.shogun.lib.repository.security.permission.UserClassPermissionRepository;
-import de.terrestris.shogun.lib.security.SecurityContextUtil;
 import de.terrestris.shogun.lib.service.BaseService;
-import java.util.List;
-import java.util.Optional;
-
+import de.terrestris.shogun.lib.service.security.provider.UserProviderService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 @Log4j2
 public class UserClassPermissionService extends BaseService<UserClassPermissionRepository, UserClassPermission> {
 
     @Autowired
-    protected SecurityContextUtil securityContextUtil;
+    protected PermissionCollectionRepository permissionCollectionRepository;
 
     @Autowired
-    protected PermissionCollectionRepository permissionCollectionRepository;
+    private UserProviderService userProviderService;
 
     /**
      * Returns all {@link UserClassPermission} for the given query arguments.
@@ -109,7 +109,7 @@ public class UserClassPermissionService extends BaseService<UserClassPermissionR
      * @param permissionCollectionType The permission to set.
      */
     public void setPermission(Class<? extends BaseEntity> clazz, PermissionCollectionType permissionCollectionType) {
-        Optional<User> activeUser = securityContextUtil.getUserBySession();
+        Optional<User> activeUser = userProviderService.getUserBySession();
 
         if (activeUser.isEmpty()) {
             throw new RuntimeException("Could not detect the logged in user.");
