+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/service/security/permission/UserInstancePermissionService.java
@@ -23,23 +23,19 @@ import de.terrestris.shogun.lib.model.security.permission.PermissionCollection;
 import de.terrestris.shogun.lib.model.security.permission.UserInstancePermission;
 import de.terrestris.shogun.lib.repository.security.permission.PermissionCollectionRepository;
 import de.terrestris.shogun.lib.repository.security.permission.UserInstancePermissionRepository;
-import de.terrestris.shogun.lib.security.SecurityContextUtil;
 import de.terrestris.shogun.lib.service.BaseService;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import lombok.extern.log4j.Log4j2;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 @Service
 @Log4j2
 public class UserInstancePermissionService extends BaseService<UserInstancePermissionRepository, UserInstancePermission> {
-
-    @Autowired
-    protected SecurityContextUtil securityContextUtil;
 
     @Autowired
     protected PermissionCollectionRepository permissionCollectionRepository;
@@ -137,23 +133,6 @@ public class UserInstancePermissionService extends BaseService<UserInstancePermi
         Optional<UserInstancePermission> userInstancePermission = this.findFor(entity, user);
 
         return getPermissionCollection(userInstancePermission);
-    }
-
-    /**
-     * Sets the given {@link PermissionCollectionType} for the given entity and the currently
-     * logged in user.
-     *
-     * @param persistedEntity The entity to set the permission for.
-     * @param permissionCollectionType The permission to set.
-     */
-    public void setPermission(BaseEntity persistedEntity, PermissionCollectionType permissionCollectionType) {
-        Optional<User> activeUser = securityContextUtil.getUserBySession();
-
-        if (activeUser.isEmpty()) {
-            throw new RuntimeException("Could not detect the logged in user.");
-        }
-
-        setPermission(persistedEntity, activeUser.get(), permissionCollectionType);
     }
 
     /**
