+++ shogun-lib/src/main/java/de/terrestris/shogun/lib/util/KeycloakUtil.java
@@ -23,8 +23,12 @@ import de.terrestris.shogun.lib.repository.UserRepository;
 import de.terrestris.shogun.lib.security.SecurityContextUtil;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
+import org.keycloak.KeycloakPrincipal;
+import org.keycloak.KeycloakSecurityContext;
 import org.keycloak.admin.client.CreatedResponseUtil;
 import org.keycloak.admin.client.resource.*;
+import org.keycloak.representations.AccessToken;
+import org.keycloak.representations.IDToken;
 import org.keycloak.representations.idm.GroupRepresentation;
 import org.keycloak.representations.idm.RoleRepresentation;
 import org.keycloak.representations.idm.UserRepresentation;
@@ -171,15 +175,6 @@ public class KeycloakUtil {
     }
 
     /**
-     * @deprecated
-     * This method was moved to the {@link SecurityContextUtil} as it receives an authentication as parameter.
-     */
-    @Deprecated
-    public String getKeycloakUserIdFromAuthentication(Authentication authentication) {
-        return securityContextUtil.getKeycloakUserIdFromAuthentication(authentication);
-    }
-
-    /**
      * Fetch user name of user from keycloak
      * @param user
      * @return
@@ -240,6 +235,33 @@ public class KeycloakUtil {
             log.trace("Full stack trace: ", e);
         }
         return roles;
+    }
+
+    /**
+     * Return keycloak user id from {@link Authentication} object
+     *   - from {@link IDToken}
+     *   - from {@link org.keycloak.Token}
+     * @param authentication The Spring security authentication
+     * @return The keycloak user id token
+     */
+    public static String getKeycloakUserIdFromAuthentication(Authentication authentication) {
+        if (authentication.getPrincipal() instanceof KeycloakPrincipal) {
+            KeycloakPrincipal<KeycloakSecurityContext> keycloakPrincipal = (KeycloakPrincipal<KeycloakSecurityContext>) authentication.getPrincipal();
+            KeycloakSecurityContext keycloakSecurityContext = keycloakPrincipal.getKeycloakSecurityContext();
+            IDToken idToken = keycloakSecurityContext.getIdToken();
+            String keycloakUserId;
+
+            if (idToken != null) {
+                keycloakUserId = idToken.getSubject();
+            } else {
+                AccessToken accessToken = keycloakSecurityContext.getToken();
+                keycloakUserId = accessToken.getSubject();
+            }
+
+            return keycloakUserId;
+        } else {
+            return null;
+        }
     }
 
 }
