+++ shogun-lib/src/test/java/de/terrestris/shogun/lib/service/security/permission/GroupClassPermissionServiceTest.java
@@ -16,23 +16,34 @@
  */
 package de.terrestris.shogun.lib.service.security.permission;
 
+import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.model.security.permission.GroupClassPermission;
 import de.terrestris.shogun.lib.repository.security.permission.GroupClassPermissionRepository;
 import de.terrestris.shogun.lib.service.BaseServiceTest;
+import de.terrestris.shogun.lib.service.security.provider.UserProviderService;
+import de.terrestris.shogun.lib.service.security.provider.keycloak.KeycloakUserProviderService;
 import org.junit.Before;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+
+import java.util.Optional;
+
+import static org.mockito.Mockito.when;
 
 public class GroupClassPermissionServiceTest extends BaseServiceTest<GroupClassPermissionService, GroupClassPermission> {
 
     @Mock
     GroupClassPermissionRepository repositoryMock;
 
+    @Mock
+    private UserProviderService userProviderService = new KeycloakUserProviderService();
+
     @InjectMocks
     GroupClassPermissionService service;
 
     @Before
     public void init() {
+        when(userProviderService.getUserBySession()).thenReturn(Optional.of(new User()));
         super.setRepository(repositoryMock);
         super.setService(service);
         super.setEntityClass(GroupClassPermission.class);
