+++ shogun-lib/src/test/java/de/terrestris/shogun/lib/service/security/permission/UserInstancePermissionServiceTest.java
@@ -16,23 +16,33 @@
  */
 package de.terrestris.shogun.lib.service.security.permission;
 
+import de.terrestris.shogun.lib.model.User;
 import de.terrestris.shogun.lib.model.security.permission.UserInstancePermission;
 import de.terrestris.shogun.lib.repository.security.permission.UserInstancePermissionRepository;
 import de.terrestris.shogun.lib.service.BaseServiceTest;
+import de.terrestris.shogun.lib.service.security.provider.keycloak.KeycloakUserProviderService;
 import org.junit.Before;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+
+import java.util.Optional;
+
+import static org.mockito.Mockito.when;
 
 public class UserInstancePermissionServiceTest extends BaseServiceTest<UserInstancePermissionService, UserInstancePermission> {
 
     @Mock
     UserInstancePermissionRepository repositoryMock;
 
+    @Mock
+    KeycloakUserProviderService userProviderService = new KeycloakUserProviderService();
+
     @InjectMocks
     UserInstancePermissionService service;
 
     @Before
     public void init() {
+        when(userProviderService.getUserBySession()).thenReturn(Optional.of(new User()));
         super.setRepository(repositoryMock);
         super.setService(service);
         super.setEntityClass(UserInstancePermission.class);
